backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pyjwt==2.6.0
python-dotenv==1.0.0
email-validator==2.2.0

